Implement raw mode

+++ line.h
@@ -5,7 +5,7 @@
 typedef char Letter;
 
 typedef struct {
-    size_t size;
+    int size;
     Letter* letters; // null terminated (not counted in size) **************************
 } Line;
 
@@ -13,7 +13,7 @@ void line_print(const Line *line);
 void line_init(Line *line);
 void line_delete(Line *line);
 
-void line_insert_letter_at(Line *line, size_t at, Letter l);
-void line_delete_letter_at(Line *line, size_t at);
+void line_insert_letter_at(Line *line, int at, Letter l);
+void line_delete_letter_at(Line *line, int at);
 
-void line_append_str(Line *line, const char* buf, size_t buflen);
+void line_append_str(Line *line, const char* buf, int buflen);
